refactor(quick-open): split ripgrep guidance
